+++ linux/rollup/http/echo-dapp/src/config.rs
@@ -10,24 +10,27 @@
 // CONDITIONS OF ANY KIND, either express or implied. See the License for the
 // specific language governing permissions and limitations under the License.
 
+#[derive(Clone, Debug, PartialEq)]
+pub struct TestConfig {
+    pub vouchers: u32,
+    pub reports: u32,
+    pub notices: u32,
+    pub reject: i32,
+    pub exception: i32,
+}
+
 /// Application configuration
 #[derive(Clone, Debug, PartialEq)]
 pub struct Config {
-    pub http_address: String,
-    pub http_port: u16,
-    pub dispatcher_address: String,
-    pub dispatcher_port: u16,
-    pub reject: i32,
+    pub rollup_http_server_address: String,
+    pub test_config: TestConfig,
 }
 
 impl Config {
-    pub fn new() -> Self {
+    pub fn new(test_config: TestConfig) -> Self {
         Self {
-            http_address: String::from("127.0.0.1"),
-            http_port: 5002,
-            dispatcher_address: String::from("127.0.0.1"),
-            dispatcher_port: 5001,
-            reject: -1,
+            rollup_http_server_address: String::from("127.0.0.1:5001"),
+            test_config,
         }
     }
 }